reordered folder-file structure
as asked in last pull request I reordered the folder structure, but I didnt add emotion css
