+++ src/modules/configurator/assets/pizza.png
