+++ src/modules/configurator/components/discount/Discount.module.css
