+++ src/modules/configurator/components/price-view/PriceView.module.css
