+++ src/modules/configurator/components/sizer/Sizer.module.css
