+++ src/modules/configurator/components/toppings/ToppingsItem.tsx
